Add a /rebase command
Let users rebase their ghstack stacks by leaving a /rebase comment
on a PR. This action is only permitted if the commenter is the
owner of the PR or a maintainer of the repo.

The action acts using the fms-bot account, it attempts to
rebase (git rebase origin/main), and if no conflicts are found, it
pushes the rebase via ghstack.

Note that the fms-bot is not the same account as the author or the PRs
The original author(s) will be preserved in the it log, but the committer
will be set to fms-bot.

If the rebase command is issued on a PR that is not the top PR of
a stack, the rebase command will only rebase the PR where the
comment was issued and those underneath it in the stack.

The python code that performs the checks for the "/land" command has
been refactored, shared checks have been separated so that they may
be reused by other commands, like the "/rebase" implemented here.

Shared checks verify that the PR is formatted according to ghstack
requirements and attempt to extract the list of PRs in the stack.

Fixes #142

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

[ghstack-poisoned]